feat: Project structure refactor
