+++ app/src/main/java/com/example/ecomonitor/domain/interfaces/IObservability.kt
@@ -1,0 +1,5 @@
+package com.example.ecomonitor.domain.interfaces
+
+interface IObservability {
+    fun log(eventName: String): Unit
+}
