+++ app/src/main/java/com/example/ecomonitor/domain/model/Measurement.kt
@@ -1,0 +1,8 @@
+package com.example.ecomonitor.domain.model
+
+import java.util.Date
+
+data class Measurement(
+    val date: Date,
+    val value: Int
+)
