+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:elevation="2dp"
     android:textAlignment="textEnd"
-    tools:context=".MainActivity"
+    tools:context=".presentation.view.MainActivity"
     tools:layout_editor_absoluteX="-66dp"
     tools:layout_editor_absoluteY="0dp">
